fix: move headers/manifest validation to Tier-2, fix run_dump name, protect forced_includes
Several CodeRabbit/Codex findings after the rebase onto main:

- InputSpec.headers/dump_manifest mutual-exclusivity was only enforced at
  runtime inside run_compare_request; moved into
  CompareRequest.validation_errors() (the Tier-2 D9 pre-flight contract
  every other rule here already follows), so a caller using
  validation_errors()/validate() alone also catches it. The now-redundant
  runtime check in service.py was removed (dead code once validate()
  raises first).

- service.run_dump's __name__/__qualname__ read as "_run_dump_uncached"
  instead of "run_dump": two nested functools.wraps() calls building the
  public facade each copy __name__ down from the wrapped function. Reset
  explicitly after construction; the extended __signature__
  (include_dependencies) was already correct.

- A --dump-manifest translation unit's forced_includes weren't protected
  from dependency-scope filtering: a TU may force-include a private
  support header alongside a public one without it being in roots or
  marked project_owned, so that header's declarations were misclassified
  as toolchain dependencies. dump_manifest_header_roots now folds in
  forced_includes too.

- A test in test_service_unit.py could pass while asserting nothing (a
  try/except ValidationError with no assertion on the non-raising path);
  fixed to always assert on the captured message.

- Two comments/docstrings describing compare's live-binary dumping as
  never filtering (and that as "out of scope") were stale: a later commit
  in this same PR made compare filter by default too, matching dump.
  Updated to describe current behavior while keeping the historical
  context of what the original comparability-gate fix covered.

Split tests/test_dumper_scoping.py's now-oversized
TestDirectlyReferencedDependencyRetention class (grew past the 2000-line
AI-readiness hard cap after rebasing in PR #653's upstream work) into its
own file, tests/test_dumper_scoping_dependency_retention.py -- no test
content changed, same 90 tests total across both files.

All fixes verified genuine by reverting each and confirming the
corresponding test fails for the right reason, then restoring it. Full
fast suite: 21410 passed, 0 failed. mypy/ruff/ai-readiness clean.
